Collapse retriever into single pathrag module
