--- PathRAG/PathRAG/prompt.py
@@ -1,3 +1,0 @@
-"""Prompt templates used by the simplified retriever."""
-
-RAG_PROMPT = """You are a helpful assistant. Use the supplied context to answer the question.\n\nContext:\n{context}\n\nQuestion: {question}\nAnswer in Markdown."""
